+++ packages/flutter_fixtures_dio/lib/flutter_fixtures_dio.dart
@@ -1,3 +1,4 @@
 library flutter_fixtures_dio;
 
 export 'src/dio_interceptor.dart';
+export 'src/recorder_interceptor.dart';