+++ packages/flutter_fixtures_recorder_dio/lib/flutter_fixtures_recorder_dio.dart
@@ -1,5 +1,0 @@
-library flutter_fixtures_recorder_dio;
-
-export 'package:flutter_fixtures_recorder/flutter_fixtures_recorder.dart';
-
-export 'src/recorder_interceptor.dart';